% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_bpc_tic.R
\name{chrom_bpc_tic}
\alias{chrom_bpc_tic}
\title{Plotting BPC and TIC}
\usage{
chrom_bpc_tic(df, pars)
}
\arguments{
\item{df}{data.frame with scantime, m/z value and intensity}

\item{pars}{plotting parameters (shiny reactive values)}
}
\value{
plotly object
}
\description{
Plotting BPC and TIC
}
