% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_xic.R
\name{chrom_xic}
\alias{chrom_xic}
\title{Plotting function for eXtracted Ion Chromatogram (XIC)}
\usage{
chrom_xic(df, pars)
}
\arguments{
\item{df}{data.frame with scantime, m/z value and intensity}

\item{pars}{plotting parameters (shiny reactive values)}
}
\description{
Plotting function for eXtracted Ion Chromatogram (XIC)
}
