% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{icst}
\alias{icst}
\title{UI elements used with insertUI and removeUI}
\format{An object of class \code{data.frame} with 31 rows and 5 columns.}
\usage{
icst
}
\description{
UI elements used with insertUI and removeUI
}
\keyword{datasets}
