% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massspectrum.R
\name{massspectrum}
\alias{massspectrum}
\title{Plotting function for single scan mass spectrum}
\usage{
massspectrum(df, pars)
}
\arguments{
\item{df}{data.frame with scantime, m/z value and intensity}

\item{pars}{plotting parameters (shiny reactive values)}
}
\description{
Plotting function for single scan mass spectrum
}
