% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{transf}
\alias{transf}
\title{Transformation functions}
\usage{
transf(df, trans)
}
\arguments{
\item{df}{data.frame with scantime, m/z value and intensity}

\item{trans}{char indicating transformation methods}
}
\value{
data frame with transformed counts
}
\description{
Transformation functions
}
