% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcms_df.R
\name{xcms_df}
\alias{xcms_df}
\title{Read-in LC-MS file (xcms) and extract raw scan data and create dataframe}
\usage{
xcms_df(raw_xcms)
}
\arguments{
\item{raw_xcms}{xcms data object}
}
\description{
Read-in LC-MS file (xcms) and extract raw scan data and create dataframe
}
