% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_bpc_tic.R
\name{chrom_bpc_tic}
\alias{chrom_bpc_tic}
\title{Plotting function for base peak and total ion chromatrogram in a single plot}
\usage{
chrom_bpc_tic(df = raw_data()[[1]], pars)
}
\arguments{
\item{df}{data.frame with scantime, m/z value and intensity}

\item{pars}{plotting parameters (shiny reactive values)}
}
\description{
Plotting function for base peak and total ion chromatrogram in a single plot
}
